fix banner to nor resurface if existing TCF prefs exist
